move has*Permission to AccessManager
